Revert "Implement macOS as a build target"
This reverts commit 1620b376.

After some initial testing - it was found that Rosetta stores it's application translation files in a temporary directory that PLRPC couldn't write to. While this is a problem, the underlying issue is with the updater, as it attempts to write a manifest file to the working directory of the application. Seeing as this can't be done, it was determined that changes to the updater should be made before macOS support be implemented.

Reopening #36.